refactor(wallet_policy): move KeyInfo into its own module
